requirements.txt
numpy==1.24.3
opencv-python==4.8.0.76
pillow==10.0.0
matplotlib==3.7.2
tensorflow-macos==2.13.0  # Apple Silicon specific
tensorflow-metal==1.0.0   # GPU acceleration for Apple Silicon
scikit-learn==1.3.0
seaborn==0.12.2
tqdm==4.65.0
jupyter==1.0.0
ipykernel==6.25.0
ipywidgets==8.1.0
streamlit